Ready to submit. Produces several mandelbrot images with different sizes.

+++ pa3/kernel.h
@@ -2,6 +2,6 @@
 #define KERNEL_H_
 
 // This is the declaration of the function that will execute on the GPU.
-__global__ void mandel(int *, int, int, int, int, int, int, int);
+__global__ void mandel(int *, int, float, float, float, float, int, int);
 
 #endif // KERNEL_H_
